+++ build_runner_core/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 6.1.5
+
+- Allow build version `1.6.x`.
+
 ## 6.1.4
 
 - Allow build_config version `'>=0.4.1 <0.4.6'`.